Support airplane-mode indicator switch
Starting on Linux 4.6 the RFKill subsystem provides means for userspace
to control an airplane-mode indicator, which platform drivers use as a
trigger to drive airplane-mode LEDs that might be present on the
machine.

Mutually exclusive access to this control knob is enforced by the
kernel, by requiring /dev/rfkill to be held open while controlling the
airplane-mode indicator state.